Fix zero-size struct member usage
Apply zero-sized-members.patch:

Description: Fix zero-size struct member usage
Author: Guillem Jover <gjover@sipwise.com>

These zero-members are not at the end of the struct, which is an old
convention/extension to implement flexible member arrays. In this case
it appears to be used as a replacement for unions.

Replace this kind of usage with actual unions, to preserve the ABI.